multiprocess: Optimize process memory sharing
Multiple processes need to access the same large arrays when solving PDE.  In
order to efficiently transmit data between the processes, we use np.memmap
rather than mp.Array.  This avoids sending data over pipes, which is quite
costly.  In addition, we further reduce the amount of data to be sent each time
step by only communicating the updated time, rather than sending the system via
pickle along with the input data.  Still need to do a bit of cleanup but we
have nice profiling results up to the number of cores.